New graph - area graph
Connecticut for 19 April 2020

Each area's size is related to cumulative affected population.

Steepness of each curve related to growth rates.

Time index is left to right (X axis) and population
(in thousands) on Y axis.

Thus 17,550 is magnitude of entire State's confirmed
cases population.

Adjacent areas (dark blue, green and so on) are
taken similarly to a stacked-bar type of graph.

	new file:   connecticut_covid_19_apr_a.jpg

On branch temp-rvw-