Updates through 9 June, rel. complete.
This should be a solid update, with most
expected edits in place for this series.

Likely a candidate to merge all the way
back into master, and soon.

The State changed its reporting format,
several days back.  The present work
incorporates that, fully (unless there's
been an oversight).

	modified:   covid-19_progression_math.gnumeric

On branch tmp-conf-probs-aa-